requirements.txt
django==3.2.15
djangorestframework==3.14.0